# worker/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bluby API",
    "version": "1.0.0",
    "description": "Cloudflare Worker backend for the BlubyAI application"
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Create login",
        "description": "Create operation for /auth/login",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Create register",
        "description": "Create operation for /auth/register",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Create logout",
        "description": "Create operation for /auth/logout",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/profiles/{userId}": {
      "get": {
        "summary": "Get a user profile by ID",
        "description": "Retrieve operation for /v1/profiles/:userId",
        "tags": [
          "profiles"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "summary": "Update a user profile",
        "description": "Partially update operation for /v1/profiles/:userId",
        "tags": [
          "profiles"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/meals/": {
      "get": {
        "summary": "Get meals",
        "description": "Retrieve operation for /v1/meals/",
        "tags": [
          "meals"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create meals",
        "description": "Create operation for /v1/meals/",
        "tags": [
          "meals"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/friends/": {
      "get": {
        "summary": "Get friends",
        "description": "Retrieve operation for /v1/friends/",
        "tags": [
          "friends"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/friends/request/{userId}": {
      "post": {
        "summary": "Send friend request",
        "description": "Create operation for /v1/friends/request/:userId",
        "tags": [
          "friends"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/friends/requests": {
      "get": {
        "summary": "Get friend requests",
        "description": "Retrieve operation for /v1/friends/requests",
        "tags": [
          "friends"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/groups/": {
      "get": {
        "summary": "Get groups",
        "description": "Retrieve operation for /v1/groups/",
        "tags": [
          "groups"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create groups",
        "description": "Create operation for /v1/groups/",
        "tags": [
          "groups"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/groups/{groupId}": {
      "get": {
        "summary": "Get specific group",
        "description": "Retrieve operation for /v1/groups/:groupId",
        "tags": [
          "groups"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/ingredients/": {
      "get": {
        "summary": "Get ingredients",
        "description": "Retrieve operation for /v1/ingredients/",
        "tags": [
          "ingredients"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create ingredients",
        "description": "Create operation for /v1/ingredients/",
        "tags": [
          "ingredients"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/ingredients/{ingredientId}": {
      "get": {
        "summary": "Get specific ingredient",
        "description": "Retrieve operation for /v1/ingredients/:ingredientId",
        "tags": [
          "ingredients"
        ],
        "parameters": [
          {
            "name": "ingredientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/products/": {
      "get": {
        "summary": "Get products",
        "description": "Retrieve operation for /v1/products/",
        "tags": [
          "products"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create products",
        "description": "Create operation for /v1/products/",
        "tags": [
          "products"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/products/{productId}": {
      "get": {
        "summary": "Get specific product",
        "description": "Retrieve operation for /v1/products/:productId",
        "tags": [
          "products"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/preferences/": {
      "get": {
        "summary": "Get preferences",
        "description": "Retrieve operation for /v1/preferences/",
        "tags": [
          "preferences"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update preferences",
        "description": "Update operation for /v1/preferences/",
        "tags": [
          "preferences"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create preferences",
        "description": "Create operation for /v1/preferences/",
        "tags": [
          "preferences"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/recipes/": {
      "get": {
        "summary": "Get recipes",
        "description": "Retrieve operation for /v1/recipes/",
        "tags": [
          "recipes"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create recipes",
        "description": "Create operation for /v1/recipes/",
        "tags": [
          "recipes"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/recipes/{recipeId}": {
      "get": {
        "summary": "Get specific recipe",
        "description": "Retrieve operation for /v1/recipes/:recipeId",
        "tags": [
          "recipes"
        ],
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/food_scan/scan": {
      "post": {
        "summary": "Scan food image",
        "description": "Create operation for /food_scan/scan",
        "tags": [
          "foodScan"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProfileDetails": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "allow_contact_search": {
            "type": "boolean"
          }
        }
      },
      "SuccessResponseProfileDetails": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "body": {
            "$ref": "#/components/schemas/ProfileDetails"
          }
        }
      }
    }
  }
}