# worker/src/routes/groups.ts
import { Hono } from 'hono';
import { HTTPException } from 'hono/http-exception';
import { authMiddleware, AuthUser } from '../middleware/auth';
import { Env } from '../auth';

const groups = new Hono<{ Bindings: Env; Variables: { user: AuthUser } }>();

// Get user groups
groups.get('/', authMiddleware, async (c) => {
  // Mock implementation
  return c.json([]);
});

// Create a new group
groups.post('/', authMiddleware, async (c) => {
  // Mock implementation
  return c.json({});
});

// Get specific group
groups.get('/:groupId', authMiddleware, async (c) => {
  const groupId = c.req.param('groupId');
  
  try {
    const groupIdNum = parseInt(groupId);
    if (isNaN(groupIdNum)) {
      throw new HTTPException(422, { 
        message: JSON.stringify({ 
          detail: [{ loc: ['path', 'groupId'], msg: 'Invalid group ID', type: 'value_error' }] 
        })
      });
    }

    // Mock implementation
    return c.json({});
  } catch (error) {
    if (error instanceof HTTPException) {
      throw error;
    }
    const errorMessage = error instanceof Error ? error.message : 'Unknown error';
    throw new HTTPException(500, { 
      message: JSON.stringify({ 
        detail: [{ loc: [], msg: errorMessage, type: 'server_error' }] 
      })
    });
  }
});

export default groups;